+++ src/layers/hillshade.tsx
@@ -1,7 +1,8 @@
-import { Layer, Source } from "react-map-gl/maplibre"
-import { demSource, elevationEncoding, elevationScheme, maxContourZoom } from "./contours"
+import { demSource, elevationEncoding, elevationScheme, maxContourZoom } from "./demSource"
 import React from "react"
-import { OverlayDefinition } from "./layerDefinition"
+import { OverlayDefinition } from "./config"
+import Source from '../map/Source'
+import Layer from '../map/Layer'
 
 export default {
     id: 'hillshade',
@@ -9,9 +10,19 @@ export default {
     description: 'Hillshade tiles for the terrain',
     type: 'overlay',
     cacheable: false,
-    source: <Source key='hillshade' id='hillshade' tileSize={256} tiles={[demSource.sharedDemProtocolUrl]} maxzoom={maxContourZoom} type='raster-dem' encoding={elevationEncoding} scheme={elevationScheme}>
-        <Layer id="hillshade" type='hillshade' source='terrain' paint={{
-            "hillshade-exaggeration": 0.1,
+    source: <React.Fragment key='hillshade'>
+        <Source id='hillshade' spec={{
+            tileSize: 256,
+            tiles: [demSource.sharedDemProtocolUrl],
+            maxzoom: maxContourZoom,
+            type: 'raster-dem',
+            encoding: elevationEncoding,
+            scheme: elevationScheme
         }} />
-    </Source>
+        <Layer layer={{
+            id: 'hillshade', type: 'hillshade', source: 'terrain', paint: {
+                "hillshade-exaggeration": 0.1,
+            }
+        }} />
+    </React.Fragment>
 } as OverlayDefinition
