+++ src/layers/terrain.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import Source from "../map/Source"
-import { demSource, elevationEncoding, elevationScheme, maxContourZoom } from "./contours"
+import { demSource, elevationEncoding, elevationScheme, maxContourZoom } from "./demSource"
 
 export default function Terrain() {
     return <Source id="dem" spec={{
